Add css styling for Grocery List component.
