+++ src/pages/GroceryList/GroceryList.css
@@ -1,0 +1,41 @@
+.grocery-list-wrapper {
+  display: flex;
+  justify-content: center;
+}
+
+.list-items {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  /* Border box */
+  background: var(--cream_opaque);
+  padding: 1rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 0.25rem 0.375rem var(--shadow_black);
+  width: fit-content;
+}
+
+.list-item {
+  display: flex;
+  justify-content: space-between;
+  gap: 1rem;
+  width: 100%; /* Takes full width of the wrapper */
+  margin-bottom: 0.5rem;
+  align-items: center;
+}
+
+.remove-btn {
+  cursor: pointer;
+  background: none;
+  border: none;
+  transition: all 0.2s ease;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  color: var(--error_red);
+  flex-shrink: 0;
+}
+
+.remove-btn:hover {
+  transform: translateY(-1px);
+}
